Move minting logic out of GratitudeForm
